
\docType{data}
\name{beta.emp}
\alias{beta.emp}
\title{beta.emp}
\format{A dataframe containing methylaton beta values for 62 tumor and normal samples.}
\source{
X. Zheng, N. Zhang, H.J. Wu and H. Wu, Estimating and accounting for tumor purity in the analysis of DNA methylation data from cancer studies. \emph{Genome biology}, accepted. 
}
\usage{
beta.emp
}
\description{
An example data set for InfiniumClust and InfiniumPurify.
}

