\name{CancerTypeAbbr}
\alias{CancerTypeAbbr}
\title{
Print abbreviations of cancer types with known iDMCs. 
}
\description{
Print tumor types and their abbreviations with known informative DMCs.  
}
\usage{
CancerTypeAbbr()
}

\arguments{
  None.
}

\references{
X. Zheng, N. Zhang, H.J. Wu and H. Wu, Estimating and accounting for tumor purity in the analysis of DNA methylation data from cancer studies. \emph{Genome biology} \bold{18}, 17. 
}
\author{
Xiaoqi Zheng \email{xqzheng@shnu.edu.cn}.
}

\examples{
data(abbr)
CancerTypeAbbr()
}