
\docType{data}
\name{iDMC}
\alias{iDMC}
\title{iDMC}
\format{A list containing informative Differential methylation CpG sites (iDMC) and their average methylation levels in tumor and normal samples.}
\source{
X. Zheng, N. Zhang, H.J. Wu and H. Wu, Estimating and accounting for tumor purity in the analysis of DNA methylation data from cancer studies. \emph{Genome biology} \bold{18}, 17. 
}
\usage{
iDMC
}
\description{
This data set lists pre-selected iDMCs for all TCGA cancer types.
}
\keyword{datasets}

