
\docType{data}
\name{abbr}
\alias{abbr}
\title{abbr}
\format{A dataframe containing names and abbreviations for all TCGA cancer types.}
\source{
X. Zheng, N. Zhang, H.J. Wu and H. Wu, Estimating and accounting for tumor purity in the analysis of DNA methylation data from cancer studies. \emph{Genome biology} \bold{18}, 17. 
}
\usage{
abbr
}
\description{
This data set lists abbreviations for all TCGA cancer types.
}
\keyword{datasets}
