
\docType{data}
\name{beta.emp}
\alias{beta.emp}
\title{beta.emp}
\format{A dataframe containing methylaton beta values for 61 tumor and normal samples.}
\source{
X. Zheng, N. Zhang, H.J. Wu and H. Wu, Estimating and accounting for tumor purity in the analysis of DNA methylation data from cancer studies. \emph{Genome biology} \bold{18}, 17. 

W. Zhang, Z. Li, N. Wei, H. Wu and X. Zheng, Detection of differentially methylated CpG sites between tumor samples with uneven tumor purities, in submission.
}
\usage{
beta.emp
}
\description{
An example data set for InfiniumDM, InfiniumClust and InfiniumPurify.
}
\keyword{datasets}
